Create create-extent and get-spatial-reference helpers Update layers-info-route-handler and relationships-info-route-handler to use new helpers Update tests to handle new helpers rename server-info-route-handler.relationshipInfo variables for clarity
