+++ lib/layers-info-route-handler.js
@@ -1,15 +1,13 @@
 const _ = require('lodash')
-const { calculateBounds } = require('@terraformer/spatial')
 const {
-  getCollectionCrs,
+  calculateExtent,
   getGeometryTypeFromGeojson,
+  getSpatialReference,
   normalizeExtent,
-  normalizeSpatialReference,
   normalizeInputData
 } = require('./helpers')
 const { computeFieldObject } = require('./field')
 const { layerMetadata: layerMetadataDefaults, renderers: rendererDefaults } = require('./defaults')
-const debug = process.env.KOOP_LOG_LEVEL === 'debug' || process.env.LOG_LEVEL === 'debug'
 
 module.exports = function layersMetadata (providerResponse, queryParams = {}) {
   const { layers: layersInput, tables: tablesInput } = normalizeInputData(providerResponse)
@@ -84,37 +82,4 @@ function formatResponse (geojson = {}, options = {}) {
   return _.defaults(json, {
     capabilities: extract ? `${layerMetadataDefaults.capabilities},Extract` : layerMetadataDefaults.capabilities
   }, layerMetadataDefaults)
-}
-
-function calculateExtent ({ isLayer, geojson, spatialReference }) {
-  if (!isLayer) {
-    return
-  }
-
-  try {
-    const bounds = calculateBounds(geojson)
-    return normalizeExtent(bounds, spatialReference)
-  } catch (error) {
-    if (debug) {
-      console.log(`Could not calculate extent from data: ${error.message}`)
-    }
-  }
-}
-
-function getSpatialReference (geojson, {
-  inputCrs,
-  sourceSR
-}) {
-  if (!inputCrs && !sourceSR && _.isEmpty(geojson)) return
-  const spatialReference = inputCrs || sourceSR || getCollectionCrs(geojson)
-
-  if (!spatialReference) return
-
-  const { latestWkid, wkid, wkt } = normalizeSpatialReference(spatialReference)
-
-  if (wkid) {
-    return { wkid, latestWkid }
-  }
-
-  return { wkt }
 }
