+++ lib/server-info-route-handler.js
@@ -107,15 +107,15 @@ function formatInfo (json = {}, defaultId, type) {
   }
 }
 
-function relationshipInfo (json = {}, defaultId) {
+function relationshipInfo (json = {}, relationshipIndex) {
   const {
     id,
     name
   } = json
 
-  const defaultName = `Relationship_${id || defaultId}`
+  const defaultName = `Relationship_${id || relationshipIndex}`
   return {
-    id: id || defaultId,
+    id: id || relationshipIndex,
     name: name || defaultName
   }
 }
